[BuildOutput] Avoid unneeded UI refreshes and cancel tree searching correctly
A big improvement in the individual node's searching, as the method that does
so now receives a CancellationToken which it checks for every pass of the
children traversal loop, thus allowing termination as soon as the operation is
canceled (instead of just waiting for the whole tree to be searched, as it
was doing before).

Also, if searches are canceled, it means we never finished getting any data
from it, so no need to refresh rows by raising NodeChanged event. This avoids
a lot of extra row re-calculations also.