[MacPlatform] Add HttpClient options panel
Added a new options panel that can be used to switch between
NSUrlSession and a managed (mono) implementation when using the
HttpClient. This options panel is available from Other - Network
and has a combo box that can switch between the two http client
handler implementations. NSUrlSession is used by default. Currently
this does not affect when http client handler is used by the IDE.